+++ src/comms/driver_station/commands/disable_command.rs
@@ -1,8 +1,8 @@
 use crate::comms::driver_station::DriverStationController;
 use crate::comms::driver_station::SubsystemIdentifier;
-use crate::comms::reading::Command;
+use crate::comms::parsing::Command;
 use crate::drive_train::DriveTrainCommand;
-use crate::comms::reading::CommandParser;
+use crate::comms::parsing::CommandParser;
 use crate::logging::log_data::LogData;
 use crate::comms::get_wrong_arg_count_log;
