+++ src/comms/driver_station/commands/kill_command.rs
@@ -1,7 +1,7 @@
 use crate::drive_train::DriveTrainCommand;
 use crate::comms::driver_station::DriverStationController;
-use crate::comms::reading::Command;
-use crate::comms::reading::CommandParser;
+use crate::comms::parsing::Command;
+use crate::comms::parsing::CommandParser;
 use crate::logging::log_data::LogData;
 use crate::comms::get_wrong_arg_count_log;
